backend/requirements.txt
fastapi
uvicorn
pydantic
mistralai
plotly
kaleido
python-dotenv
